feat: filter unresolved comments
Unresolved status is on comments but we want to show only comments that
are part of a "thread" that is not solved.

In order to do that we need to manually build thread of comments.
Comment has a "in_response_to" field we can use to get their parents and
build the thread thanks to that. We also need to sort comments by date
to properly build threads.
That is because to build thread we might need to get the parent of a
parent (via in_response_to), but if parents are not properly ordered by
date we might not find it.

Unresolved of a thread is unresolved of the last comment of a thread.